Moved tournament functionality to external component
